+++ lib/avromatic/model/messaging.rb
@@ -29,14 +29,14 @@ module Avromatic
       end
       include Encode
 
-      # This module provides methods to deserialize an Avro-encoded value and
+      # This module provides methods to message_decode an Avro-encoded value and
       # an optional Avro-encoded key as a new model instance.
       module Decode
 
         # If two arguments are specified then the first is interpreted as the
         # message key and the second is the message value. If there is only one
         # arg then it is used as the message value.
-        def deserialize(*args)
+        def message_decode(*args)
           message_key, message_value = args.size > 1 ? args : [nil, args.first]
           key_attributes = message_key && messaging.decode(message_key, schema_name: key_avro_schema.fullname)
           value_attributes = messaging.decode(message_value, schema_name: avro_schema.fullname)